mindmate/requirements.txt
# MindMate - Mental Health Multi-Agent System
# Core Dependencies

# Google ADK and Gemini
google-adk>=0.3.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.70.0

# Async and HTTP
aiohttp>=3.9.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.32.0

# MCP Protocol
mcp>=1.0.0

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
structlog>=24.4.0

# Data handling
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Storage
aiosqlite>=0.20.0
chromadb>=0.5.0

# Utilities
python-dotenv>=1.0.0
rich>=13.9.0
tenacity>=9.0.0
numpy>=1.26.0

# Evaluation
scikit-learn>=1.5.0
nltk>=3.9.0
textblob>=0.18.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

